fix: revert 'forecast stake distribution hack' introduced for testing
  It is not okay to change production behavior to accomodate tests; this is dangerous and will send even send a wrong test signal. We should test the code that ends up in production.

  Note that I've kept some of the ideas behind the hack to make the stake distribution resolution a bit more flexible while remaining correct. This is mostly just deferring the error to later; but no longer tries to return a placeholder stake distribution when not found.

+++ crates/amaru-kernel/src/cardano/era_history.rs
@@ -591,7 +591,7 @@ impl EraHistory {
             if era.contains_epoch_unchecked_horizon(&epoch) {
                 let epochs_elapsed = epoch - era.start.epoch;
                 let offset = era.start.slot;
-                let slots_elapsed = epochs_elapsed * era.params.epoch_size_slots;
+                let slots_elapsed = u64::from(epochs_elapsed) * era.params.epoch_size_slots;
                 let start = offset.offset_by(slots_elapsed);
                 let end = offset.offset_by(era.params.epoch_size_slots + slots_elapsed);
                 return Ok(EpochEraBounds { start, end: era.end.as_ref().map(|_| end) });
